Parandasin projekti ja koodi struktuuri veid
