+++ Projekt/Vanad versioonid/prototyyp1.py
